phoenix_bot/requirements.txt
python-telegram-bot==20.8
asyncpg==0.29.0
redis==5.0.1
celery==5.3.4
sqlalchemy==2.0.23
alembic==1.13.1
pydantic==2.5.0
structlog==24.1.0
prometheus-client==0.19.0
psycopg2-binary==2.9.9
dependency-injector==4.41.0
aioredis==2.0.1
python-dotenv==1.0.0